requirements.txt
flask
stripe
python-telegram-bot
python-dotenv
apscheduler
pytz
